rename fixed hour tariff
