+++ packages/modules/electricity_tariffs/fixed_hours/tariff.py
@@ -3,7 +3,7 @@ import logging
 import datetime
 import time
 
-from modules.electricity_tariffs.fixedhours.config import FixedHoursTariff, FixedHoursTariffConfiguration
+from modules.electricity_tariffs.fixed_hours.config import FixedHoursTariff, FixedHoursTariffConfiguration
 from modules.common.abstract_device import DeviceDescriptor
 from modules.common.component_state import TariffState
 # from modules.common.configurable_tariff import ConfigurableTariff
